[Patch] Switch to Chiron Sans HK (#28)
* 💈 patch: switch to Chiron Sans HK for network optimization

Currently, the use of Google Fonts require loading segmented font files in batch, affecting the loading time and leads to consuming more bandwidth units eventually. By switching to Chiron Sans HK delivered from Cloudflare POPs at jsdelivr.net, it brings a significant reduction on loading time, as well as a smaller file size but with all the font weights.

Note: This is a preliminary commit to remove Google Fonts (with 2 font weights only) in favor of Chiron Sans HK (with all generic font weights), none of the UI/UX design is altered. It is necessary to change the existing font weights to adapt to the new font files with full range of font weights.

* 🔨 patch: add back the missing font 'Oswald'

* ⚙️ patch: add back enhancement on preload/preconnect to assets

* 🧹 patch: remove unused font 'Chivo' from Google Fonts

* 🩹 patch: exact match on the font family name